+++ src/eval/requirement/generic/generic.cpp
@@ -33,7 +33,7 @@ namespace EvaluationRequirement
 {
 
 GenericBase::GenericBase(const std::string& name, const std::string& short_name, const std::string& group_name,
-                               float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                               double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
     : ProbabilityBase(name, short_name, group_name, prob, prob_check_type, false, eval_man)
 {
 }
@@ -57,7 +57,7 @@ std::string GenericBase::valueNamePlural() const
 
 
 GenericInteger::GenericInteger(const std::string& name, const std::string& short_name, const std::string& group_name,
-                       float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
+                       double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man)
     : GenericBase(name, short_name, group_name, prob, prob_check_type, eval_man)
 {
 }
@@ -246,7 +246,7 @@ std::shared_ptr<EvaluationRequirementResult::Single> GenericInteger::evaluate (
 
 
 GenericDouble::GenericDouble(const std::string& name, const std::string& short_name, const std::string& group_name,
-                               float prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man)
+                               double prob, COMPARISON_TYPE prob_check_type, double threshold, EvaluationManager& eval_man)
     : GenericBase(name, short_name, group_name, prob, prob_check_type, eval_man), threshold_(threshold)
 {
 }
