+++ src/eval/requirement/position/across.cpp
@@ -36,7 +36,7 @@ namespace EvaluationRequirement
 {
 
 PositionAcross::PositionAcross(const std::string& name, const std::string& short_name, const std::string& group_name,
-                               float prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
+                               double prob, COMPARISON_TYPE prob_check_type, EvaluationManager& eval_man,
                                float max_abs_value)
     : ProbabilityBase(name, short_name, group_name, prob, prob_check_type, false, eval_man),
       max_abs_value_(max_abs_value)
